python/generate_traces/pyproject.toml
[tool.poetry]
name = "generate-traces"
version = "0.1.0"
description = ""
authors = ["Hardik Choksi <hardik.choksi@middleware.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opentelemetry-api = "^1.26.0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-exporter-otlp = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
